ATP added

+++ plugin.sportsview/leaguenamehelper/Tennis/ATP_World_Tour_namehelper.py
@@ -1,4 +1,5 @@
-# FILE NAMES MUST CONTAIN THE TYPE OF RACE (FP1, FP2, FP3, QUALIFYING, RACE, SPRINT, GRAND PRIX) AND THE ROUND
+# FILE NAMES MUST CONTAIN THE DATE, THE NAME OF THE TOURNAMENT, AND THE LAST NAME OF THE PLAYERS WITH "vs" BETWEEN THEM
+# FOR EXAMPLE: "ATP.World.Tour.2023-01-02.Adelaide.International.1.Draper.vs.Kwon.mkv" 
 
 # IMPORTS
 # region
@@ -13,13 +14,70 @@ import shutil
 from dateutil import parser
 # endregion
 
-# Formula1Helper class
+# ATPWorldTourHelper class                                               
 # region
-class Formula1Helper:
+class ATPWorldTourHelper:                                          
     def __init__(self):
         self.my_matches_seasons_instance = None
 # endregion
-    
+
+    # Events Dictionary
+    # region
+    event_mapping = {
+        "Adelaide International 1",
+        "Adelaide International 2",
+        "Maharashtra Open",
+        "ASB Classic",
+        "Australian Open",
+        "Cordoba Open",
+        "Open Sud de France",
+        "Dallas Open",
+        "Argentina Open",
+        "Delray Beach Open",
+        "ABN AMRO World Tennis Tournament",
+        "Qatar Exxon Mobil Open",
+        "Open 13",
+        "Rio Open",
+        "Dubai Tennis Championships",
+        "Chile Open",
+        "Abierto Mexicano",
+        "BNP Paribas Open",
+        "Miami Open",
+        "Estoril Open",
+        "U.S. Mens Clay Court Championship",
+        "Grand Prix Hassan II",
+        "Monte Carlo Masters",
+        "Barcelona Open",
+        "Srpska Open",
+        "BMW Open",
+        "Mutual Madrid Open",
+        "International BNL ditalia",
+        "Geneva Open",
+        "Lyon Open",
+        "French Open",
+        "Rosmalen Grass Court Championships",
+        "Stuttgart Open",
+        "Halle Open",
+        "Queens Club Championships",
+        "Mallorca Open",
+        "Eastbourne International",
+        "Wimbledon",
+        "Nordea Open",
+        "Suisse Open Gstaad",
+        "Hall of Fame Championships",
+        "Atlanta Open",
+        "European Open",
+        "Croatian Open",
+        "Generali Open",
+        "Citi Open",
+        "Los Cabos Open",
+        "Canadian Open",
+        "Western and Southern Financial Group Masters",
+        "US Open",
+        #### ADD MORE EVENTS HERE ####
+    }
+    # endregion
+
     # entry_method method
     # region
     def entry_method(self, focused_season_name, available_matches, id_league, events, my_matches_seasons_instance):
@@ -40,9 +98,9 @@ class Formula1Helper:
         self.get_team_ids_and_round(available_matches, events)
     # endregion
 
-    # get_team_ids_and_round method (Actually getting the RACE TYPE and ROUND for this sport)
+    # get_team_ids_and_round method                                    
     # region
-    # region (NO CHANGES REQUIRED)
+    # region 
     def get_team_ids_and_round(self, available_matches, events):
         # Iterate through the team mapping to find team IDs for the current match
         for match in available_matches:
@@ -50,105 +108,100 @@ class Formula1Helper:
             team02 = None  # Initialize a placeholder for the second team's ID
             video_path = match  # Initialize a placeholder for the video path
     # endregion
-            
-            # Get the race type
+
+        # SEARCH THROUGH THE MATCH FILENAME AND GET THE DATE 
+        # region 
+        round_number = None
+        round_match = re.search(r"\d{4}-\d{2}-\d{2}|\d{4}\s+\d{2}\s+\d{2}|\d{4}.\d{2}.\d{2}|\d{2}-\d{2}-\d{4}|\d{2}\s+\d{2}\s+\d{4}|\d{2}.\d{2}.\d{4}", match)
+
+        # IF ROUND MATCH HAS A VALUE, THEN CONVERT IT TO ROUND_NUMBER
+        if round_match:
+            round_number = (round_match.group(0))
+        else:
+            print("Date not found in match:", match)
+        # endregion
+
+        # SEARCH FILENAME FOR MATCHES IN THE EVENTS DICTIONARY
+        # region
+        for event_name_pre in ATPWorldTourHelper.event_mapping:
+            normalized_event_name = event_name_pre.lower().replace(' ', '').replace('.', '')
+            normalized_match = match.lower().replace('.', '')  
+
+            if normalized_event_name in normalized_match:
+
+                # Replace periods with spaces in the working_match variable
+                working_match, _ = os.path.splitext(match)
+                working_match = working_match.replace('.', ' ')
+
+                # Find the index where the event name appears in the working_match
+                start_index = working_match.lower().index(event_name_pre.lower())
+                
+                # Extract characters from working_match starting from the event name to the end
+                event_name = working_match[start_index:]
+
+        # endregion
+
+        # IF EVENT NAME AND ROUND NUMBER (DATE) IS NOT NONE THEN CALL THE GET_EVENT_ID METHOD
+        # region
+        if event_name is not None and round_number is not None:
+            self.get_event_id(team01, team02, round_number, available_matches, events, match, event_name)
+        else:
+            print("Some information missing in match:", match)
+        # endregion
+    # endregion 
+
+    # get_event_id method                                           
+    # region
+    def get_event_id(self, team01, team02, round_number, available_matches, events, match, event_name):
+
+        # Loop through each event in the events list                
+        # region
+        for event in events:
+        # endregion
+
             # region
-            race_type = None
-            RACE = re.search(r"RACE|Grand Prix", match)
-            if RACE is not None:
-                race_type = "RACE"
-            else:
-                FP1 = re.search(r"FP1|FP01|Free Practice One|Free.Practice.One", match)
-                if FP1 is not None:
-                    race_type = "FP1"
-                else:
-                    FP2 = re.search(r"FP2|FP02|Free Practice Two|Free.Practice.Two", match)
-                    if FP2 is not None:
-                        race_type = "FP2"
-                    else:
-                        FP3 = re.search(r"FP3|FP03|Free Practice Three|Free.Practice.Three", match)
-                        if FP3 is not None:
-                            race_type = "FP3"
-                        else:
-                            QUALY = re.search(r"QUALY|Qualy|Qualifying|qualifying", match)
-                            if QUALY is not None:
-                                race_type = "QUALY"                            
-                            else:
-                                SPRINT = re.search(r"SPRINT|Sprint", match)
-                                if SPRINT is not None:
-                                    race_type = "SPRINT"                                
-                                else:
-                                    print("Race type not found in match:", match)
+            event_name_comp = event["strEvent"]
+            event_round_str = event["dateEvent"]
             # endregion
 
-            round_number_with_words = None
-            round_match = re.search(r"Round (\d+)|Round(\d+)|R\d{2}", match)
-            if round_match:
-                round_number_with_words = (round_match.group(0))
-                round_number = int(re.sub(r'\D|^0+', '', round_number_with_words))
-                print("Round number:", round_number)   
-            else:
-                print("Round number not found in match:", match)
+            # Check if either home and away teams match team01 and team02, or vice versa (USE IF A TEAM AND DATE ORGANIZED SPORT)
+            # region
+            match_found = False
 
-            if round_number is not None:
-                self.get_event_id(team01, team02, round_number, available_matches, events, match, race_type)
-            else:
-                print("Some information missing in match:", match)
-    # endregion
+            if event_name.lower() == event_name_comp.lower():
+                print("event_name and event_name_comp match")
+                match_found = True
+                break  # Exit the loop when a match is found
+            # endregion
 
-    # get_event_id method
-    # region
-    def get_event_id(self, team01, team02, round_number, available_matches, events, match, race_type):
+        # region   
+        if match_found:
+            try:
+                # Parse the round_number into a datetime object
+                parsed_round_date_time = parser.parse(round_number)  
+                parsed_round = parsed_round_date_time.date()
 
-        # Loop through each event in the events list
-        for event in events:
-            event_round = int(event["intRound"])
+                # Convert event_round_str to a datetime object
+                event_round = parser.parse(event_round_str)
 
-            # is Grand Prix or Practice 1 or Practice 2 or Practice 3 or Qualifying or Sprint is strEvent?
-            event_type_full = event["strEvent"]
-
-            # Iterate through the team mapping to find team IDs for the current match
-            FP1 = re.search(r"Practice 1", event_type_full)
-            if FP1 is not None:
-                event_type = "FP1"
-            else:
-                FP2 = re.search(r"Practice 2", event_type_full)
-                if FP2 is not None:
-                    event_type = "FP2"
-                else:
-                    FP3 = re.search(r"Practice 3", event_type_full)
-                    if FP3 is not None:
-                        event_type = "FP3"
-                    else:
-                        QUALY = re.search(r"Qualifying", event_type_full)
-                        if QUALY is not None:
-                            event_type = "QUALY"                            
-                        else:
-                            SPRINT = re.search(r"Sprint", event_type_full)
-                            if SPRINT is not None:
-                                event_type = "SPRINT"                                
-                            else:
-                                RACE = re.search(r"Grand Prix", event_type_full)
-                                if RACE is not None:
-                                    event_type = "RACE"
-                                else:
-                                    print("Race type not found in match:", event_type_full)
-
-            if race_type == event_type:
-
-                # Check if the event's round matches the provided round_number
-                if event_round == round_number:
-                    event_id = event["idEvent"]  # Access the event ID
-                    break  # Exit the loop, a match is found
+                # Compare only dates, ignoring the time component
+                if parsed_round == event_round.date():
+                    event_id = event["idEvent"]
+            except ValueError:
+                print(f"Invalid date format: {round_number}")
         else:
             # If loop completes without a match, log a message
             print("No matching event found for teams and round.")
+        # endregion
 
         # Call the get_event_info method
-        self.get_event_info(team01, team02, round_number, available_matches, events, event_id, match)
-    # endregion          
+        # region
+        if event_id:
+            self.get_event_info(team01, team02, round_number, available_matches, events, event_id, match)
+        # endregion         
+    # endregion
 
-    # get_event_info method
+    # get_event_info method                                         
     # region
     def get_event_info(self, team01, team02, round_number, available_matches, events, event_id, match):
 
@@ -167,7 +220,7 @@ class Formula1Helper:
         response = requests.get(event_url)
         # Parse the response as JSON
         data = response.json()
-        
+
         # Check if the "events" key is present in the JSON data
         if "events" in data:
             # Extract the first event's data (assuming it's an array)
@@ -180,13 +233,11 @@ class Formula1Helper:
         else:
             # If the "events" key is not found, print an error message
             print("Event data not found for event ID:", event_id)
-            
 # endregion
 
     # Getting image info start
     # region
     def get_more_info(self, event_id, event_data, match, round_number, available_matches, events, event_url):
-        print("GET MORE INFO RUNNING")
         
         # Extract relevant event information from event_data dictionary
         team01ID = None
@@ -200,13 +251,11 @@ class Formula1Helper:
 
         # Call get_teams_info method for logos badges jerseys backgrounds
         self.get_teams_info(team01, team02, team01ID, team02ID, match, round_number, available_matches, events, event_id, event_data, event_label, event_thumbnail, event_url)
-
         # endregion
 
-    # Get teams info for logos badges jerseys backgrounds
+    # Get teams info for logos badges jerseys backgrounds (USE IF NOT A TEAM SPORT)
     # region
     def get_teams_info(self, team01, team02, team01ID, team02ID, match, round_number, available_matches, events, event_id, event_data, event_label, event_thumbnail, event_url):
-        print("GET TEAMS INFO RUNNING")
         fallback_image = xbmcvfs.translatePath("special://home/addons/plugin.sportsview/allsports/media/imagenotavailable.png")
         event_response = requests.get(event_url)
         
@@ -237,11 +286,11 @@ class Formula1Helper:
         my_matches_seasons.receive_event_data(team01, team02, round_number, available_matches, events, event_id, event_data, self.my_matches_seasons_instance, match, VERSUS, event_label, event_thumbnail)
     # endregion
 
-    # Download and cache logos badges jerseys backgrounds
+    # Download and cache logos badges jerseys backgrounds (NO CHANGES REQUIRED)
     # region
     def download_and_cache_image(self, url, save_path):
-        if not os.path.exists(save_path):
-            if url.endswith("imagenotavailable.png"):  # Check if it's a local file
+        if url and not os.path.exists(save_path):
+            if url.endswith("imagenotavailable.png"):
                 # Copy the local file to the specified save_path
                 source_path = url
                 shutil.copyfile(source_path, save_path)
@@ -256,5 +305,5 @@ class Formula1Helper:
                 else:
                     print(f"Failed to download image: {url}")
         else:
-            print(f"Image already cached: {save_path}")
+            print(f"Image already cached or URL is empty: {save_path}")
     # endregion
